split up dialog.py into a few different files
